cmd/create: Switch name of the fedora-toolbox image
This is the Toolbox-side patch to start using the new name of the
fedora-toolbox image that is used to create toolboxes. To learn more
about the change, see the discussion in Fedora dist-git[0].

Users should not notice any changes to their workflow. The only thing
that will gradually change for existing users is the shift in the image
names. Example:

From: registry.fedoraproject.org/f31/fedora-toolbox:33
To:   registry.fedoraproject.org/fedora-toolbox:33

[0] https://src.fedoraproject.org/container/fedora-toolbox/pull-request/2

https://github.com/containers/toolbox/pull/641