cmd/create: Try to avoid 'latest' tags, when looking at RepoTags
The same image ID can be referred to by multiple names. It's common
that one of the names uses the 'latest' tag, while another uses a
specific version number. It's better to resolve a short image name to
one with a specific version number, because the meaning of the tag is
clear and won't change with time.

https://github.com/containers/toolbox/pull/666