Add Arch Linux toolbox image
This commit adds an Arch Linux image, created from the upstream
base-devel branch.

It includes most of the packages that the Fedora-based images do, but
currently excludes flatpak (required for flatpak-spawn), since it adds
~690 MB of extra packages/dependencies.

The base-devel image (without any extra packages) is 729 MB, when syncing
repositories 741 MB and with all the packages (excluding flatpak) 831
MB, which should be the minimum if based on archlinux:base-devel.