src/create: Handle read-only /boot
CoreOS recently made /boot read-only[0]. This caused an issue with
starting containers because /boot was mounted only with option rslave
but missed the ro option. This caused a permission issue.

This scenario is very similar to the one with /usr on Fedora Silverblue.
The solution for this is to check mount options of the path and check if
it uses the rw option or ro and then add it to the mount options in the
--volume option in 'podman create'.

Fixes: https://github.com/coreos/fedora-coreos-tracker/issues/734

[0] https://github.com/coreos/fedora-coreos-config/commit/1de21ffa98bb22995e5b059501e1955bf52b562c

https://github.com/containers/toolbox/pull/712