cmd, pkg/podman: Add command rename
Renaming of containers has been made possible by Podman v3.0.0. Let's
advertise it in Toolbox, too.

The wrapper in pkg/podman makes use of a new helper function that parses
the output of Podman in stderr (removes the "Error: " prefix) so that it
can be appended to Toolbox's error messages.