test/system: Test logging to registries
Testing of this feature is tricky as it relies on network which is
inherently flaky. So, as part of the setup two local image registries
are created and these features can be tested against them. To prevent
the removal of the registries during testing a different root for Podman
is used. The only tested registry implementation is Docker[0].

[0] https://hub.docker.com/_/registry/

https://github.com/containers/toolbox/issues/689
https://github.com/containers/toolbox/pull/787