pkg/shell: Allow extracting the stderr even when it's shown to the user
This way the standard error stream of the spawned binaries can be
inspected to get a better understanding of the failure, while still
being shown to the user when run with the '--verbose' flag.

Unfortunately, this breaks the progress bar in 'podman pull' because
the standard error stream is no longer connected to a file descriptor
that's a terminal device.

https://github.com/containers/toolbox/issues/689
https://github.com/containers/toolbox/pull/787
https://github.com/containers/toolbox/pull/823