cmd/create: pkg/utils: Fix wrong use of regexp.MatchString return value
The regexp.MatchString [1] API returns an error only when the regular
expression is faulty, and the boolean return value tells if a match was
found. In this case, the regular expression is baked into the code as a
string literal. So, unless there's a programmer error, it should always
be valid.

Fallout dd947016b3a9098f23b2a77a3616a4f5891ae8ab

[1] https://golang.org/pkg/regexp/#MatchString

https://github.com/containers/toolbox/pull/825