pkg/utils: Fix regular expression used to find IDs in image references
Turns out the braces do not need to be escaped.

The equivalent code in the POSIX shell implementation was:
  echo "$image" | grep "^[a-f0-9]\{6,64\}$"

There the braces had to be escaped because it was using grep(1) with
basic regular expressions (ie., without the --extended-regexp flag),
where the meta-characters ?, +, {, |, ( and ) lose their special
meaning unless they are escaped.

However, that was grep(1), and this is Go's regexp package.

Fallout from dd947016b3a9098f23b2a77a3616a4f5891ae8ab

https://github.com/containers/toolbox/pull/825