Added error info on non 0 and unknown exit code
When a command is executed with toolbox run and it returns a non
zero exit code, if that exit code is not handled it is just ignored.

This prevents users to identify errors when executing commands in
toolbox.

With this fix, the unhandled exit codes are returned as errors
to give the user a better error info.