src/cmd: Remove enter-related behaviour from runCommand
This is a major surgery that makes the footprint of runCommand()
smaller by removing the CLI related functionality from it and
moving it to its callers. This move causes some duplication but
that can be solved in the future by introducing a CLI-specific
abstraction.

A new error for non-existent containers had to be added to make
it possible to handle the case from callers of the function.

The behaviour of all the affected commands ('root', enter, run)
should remain the same.

https://github.com/containers/toolbox/pull/916